Bring sage16 to near-bit-parity with original SAGE
Galaxy-by-galaxy comparison against original SAGE (sage-code/sage-model)
on mini-Millennium identified six implementation-level divergence causes
(the modular architecture itself was verified faithful). After these
fixes >=98% of galaxies are bit-identical per property at z=0 (was
4-20%), with residuals at float ULP level. Full investigation, results,
and remaining-difference analysis in docs/SAGE16-PARITY-REPORT.md.
All test tiers pass; validate-modules/check-generated/check-format/
check-docs clean.

Fix A - SAGE's literal 1.414 for sqrt(2) (dominant cause, 3e-4 offset
in every disk scale radius):
- models/sage16/modules/sage_set_disk_scale_radius/sage_set_disk_scale_radius.c:
  use 1.414 literal and double-precision arithmetic, matching SAGE
- models/sage16/modules/sage_set_disk_scale_radius/_tests/test_unit_sage_set_disk_scale_radius.c:
  reference implementation updated to match

Fix B - reionization H(z) bug (Mchar overestimated by 1/h):
- models/sage16/modules/sage_reionization/sage_reionization.c: use
  code-unit Hubble (~100, h-free) instead of 100*hubble_h
- models/sage16/modules/sage_reionization/_tests/test_unit_sage_reionization.c:
  test fixtures now set MimicConfig.Hubble

Fix C - metallicity helper precision and guards:
- models/sage16/shared/metallicity.h: double precision, SAGE-exact
  semantics (gas>0 && metals>0, cap at 1.0, never negative Z)
- models/sage16/shared/sage_agn_physics.h: metallicity local to double
- models/sage16/shared/_tests/test_unit_metallicity.c: tests updated to
  SAGE-exact contract (no epsilon cutoff, negative metals return 0)

Fix D - SAGE-exact thresholds and branch parity:
- models/sage16/modules/sage_apply_cooling/sage_apply_cooling.c: >0.0 gate
- models/sage16/modules/sage_radio_mode_heating/sage_radio_mode_heating.c:
  >0.0 gate; rheat<rcool suppression branch matches SAGE exactly
- models/sage16/modules/sage_reincorporation/sage_reincorporation.c:
  skip threshold at 0.0
- models/sage16/modules/sage_satellite_stripping/sage_satellite_stripping.c:
  double metallicity; central receives unclamped strippedGas*metallicity
  (SAGE's asymmetry), no premature float rounding
- models/sage16/shared/sage_starburst_physics.h: 1e-8 metal-yield branch
- models/sage16/modules/sage_reincorporation/_tests/test_unit_sage_reincorporation.c:
  tolerate ULP-negative metals after full drainage (SAGE behavior)

Fix E - disk-SF metal yield moved after the disk-instability chain
(SAGE applies it after check_disk_instability; this drove the metals tail):
- models/sage16/modules/sage_apply_metal_enrichment/ (new module: .c,
  module_info.yaml, README.md): applies the yield and consumes
  NewStellarMass; init-time ordering checks enforce SAGE order
- models/sage16/modules/sage_apply_star_formation_supernova/sage_apply_star_formation_supernova.c:
  yield block removed; NewStellarMass no longer zeroed here; warns if
  enrichment module missing; exact-zero skip threshold; Yield/
  FracZleaveDisk params dropped
- models/sage16/modules/sage_apply_star_formation_supernova/module_info.yaml:
  parameter list and description updated
- models/sage16/modules/sage_apply_star_formation_supernova/_tests/test_unit_sage_apply_star_formation_supernova.c:
  metal tests now exercise the apply->enrichment pipeline
- models/sage16/input/sage16_mini-millennium.yaml,
  models/sage16/input/sage16_millennium.yaml: enrichment module added to
  galaxy_physics after sage_starburst_feedback

Fix F - double-precision transport (SAGE carries these as double locals):
- models/sage16/model_properties.yaml: InfallingGas, CoolingGas, Rcool,
  CoolingLambda, NewStellarMass, SupernovaReheatedMass,
  SupernovaEjectedMass, HaloBaryonFraction, UnstableDiskGasFraction to
  type double
- src/core/core_properties.yaml: dT to double (float rounding of the
  interval perturbed every dt-scaled quantity)
- models/sage16/modules/sage_calculate_cooling_budget/sage_calculate_cooling_budget.c,
  models/sage16/modules/sage_prepare_infall_budget/sage_prepare_infall_budget.c,
  models/sage16/modules/sage_apply_infall/sage_apply_infall.c: premature
  (float) casts removed so reservoirs update with float += double
- models/sage16/shared/time_parity.h: dT comparisons against double zero

Documentation:
- docs/SAGE16-PARITY-REPORT.md (new): methodology, root causes,
  quantitative before/after, vision-compliance analysis, recommendations

+++ models/sage16/modules/sage_apply_infall/sage_apply_infall.c
@@ -64,17 +64,19 @@ int sage_apply_infall_process(struct ModuleContext *ctx, struct Halo *halos, int
   struct GalaxyData *galaxy = halos[central_idx].galaxy;
 
   // Calculate infall amount for this substep
-  double infallingGas = (double)galaxy->InfallingGas / (double)ctx->num_substeps;
+  double infallingGas = galaxy->InfallingGas / (double)ctx->num_substeps;
 
   // If the halo has lost mass, subtract baryons from the ejected mass first, then the hot gas
+  // (SAGE parity: add_infall_to_hot uses a float metallicity but double products,
+  // accumulated into the float reservoirs without intermediate float rounding)
   if (infallingGas < 0.0 && galaxy->EjectedGas > 0.0f) {
     const float metallicity = mimic_get_metallicity(galaxy->EjectedGas, galaxy->MetalsEjectedGas);
-    galaxy->MetalsEjectedGas += (float)(infallingGas * metallicity);
+    galaxy->MetalsEjectedGas += infallingGas * metallicity;
     if (galaxy->MetalsEjectedGas < 0.0f) {
       galaxy->MetalsEjectedGas = 0.0f;
     }
 
-    galaxy->EjectedGas += (float)infallingGas;
+    galaxy->EjectedGas += infallingGas;
     if (galaxy->EjectedGas < 0.0f) {
       infallingGas = galaxy->EjectedGas;
       galaxy->EjectedGas = 0.0f;
@@ -87,14 +89,14 @@ int sage_apply_infall_process(struct ModuleContext *ctx, struct Halo *halos, int
   // If the halo has lost mass, subtract hot metals mass next, then the hot gas
   if (infallingGas < 0.0 && galaxy->MetalsHotGas > 0.0f) {
     const float metallicity = mimic_get_metallicity(galaxy->HotGas, galaxy->MetalsHotGas);
-    galaxy->MetalsHotGas += (float)(infallingGas * metallicity);
+    galaxy->MetalsHotGas += infallingGas * metallicity;
     if (galaxy->MetalsHotGas < 0.0f) {
       galaxy->MetalsHotGas = 0.0f;
     }
   }
 
   // Add (subtract) the ambient (enriched) infalling gas to the central galaxy hot component
-  galaxy->HotGas += (float)infallingGas;
+  galaxy->HotGas += infallingGas;
   if (galaxy->HotGas < 0.0f) {
     galaxy->HotGas = 0.0f;
     galaxy->MetalsHotGas = 0.0f;
